+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/album/songs/Response.java
@@ -1,11 +1,7 @@
 package name.sitorhy.server.model.album.songs;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.util.ArrayList;
-
 public class Response{
-    public int code;
+    public long code;
     public long ts;
     public long start_ts;
     public String traceid;
