+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/album/songs/Singer.java
@@ -1,11 +1,11 @@
 package name.sitorhy.server.model.album.songs;
 
 public class Singer{
-    public int id;
+    public long id;
     public String mid;
     public String name;
     public String title;
-    public int type;
-    public int uin;
+    public long type;
+    public long uin;
     public String pmid;
 }
