+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/album/songs/SongList.java
@@ -2,8 +2,8 @@ package name.sitorhy.server.model.album.songs;
 
 public class SongList{
     public SongInfo songInfo;
-    public int listenCount;
+    public long listenCount;
     public String uploadTime;
-    public int isThemeSong;
+    public long isThemeSong;
     public String teamStr;
 }
