+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/fav/albums/Albumlist.java
@@ -3,16 +3,16 @@ package name.sitorhy.server.model.my.fav.albums;
 import java.util.ArrayList;
 
 public class Albumlist{
-    public int albumid;
+    public long albumid;
     public String albummid;
     public String albumname;
-    public int singerid;
+    public long singerid;
     public String singermid;
     public String singername;
     public ArrayList<Singer> singer;
     public String pic;
-    public int songnum;
-    public int pubtime;
-    public int status;
-    public int ordertime;
+    public long songnum;
+    public long pubtime;
+    public long status;
+    public long ordertime;
 }
