+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/fav/albums/Data.java
@@ -3,7 +3,7 @@ package name.sitorhy.server.model.my.fav.albums;
 import java.util.ArrayList;
 
 public class Data{
-    public int totalalbum;
-    public int has_more;
+    public long totalalbum;
+    public long has_more;
     public ArrayList<Albumlist> albumlist;
 }
