+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/fav/albums/Response.java
@@ -1,10 +1,8 @@
 package name.sitorhy.server.model.my.fav.albums;
 
-import java.util.ArrayList;
-
 public class Response{
-    public int code;
-    public int subcode;
+    public long code;
+    public long subcode;
     public String msg;
     public Data data;
 }
