+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/fav/playlists/Data.java
@@ -3,7 +3,7 @@ package name.sitorhy.server.model.my.fav.playlists;
 import java.util.ArrayList;
 
 public class Data{
-    public int totaldiss;
-    public int has_more;
+    public long totaldiss;
+    public long has_more;
     public ArrayList<Cdlist> cdlist;
 }
