+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/Backpic.java
@@ -2,6 +2,6 @@ package name.sitorhy.server.model.my.playlists;
 
 public class Backpic{
     public String picurl;
-    public int type;
+    public long type;
     public String title;
 }
