+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/Cfinfo.java
@@ -5,5 +5,5 @@ public class Cfinfo{
     public String jumpurl;
     public String jumpkey;
     public CfinfoBykey cfinfo_bykey;
-    public int similar;
+    public long similar;
 }
