+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/Iconlist.java
@@ -1,8 +1,8 @@
 package name.sitorhy.server.model.my.playlists;
 
 public class Iconlist{
-    public int width;
-    public int height;
+    public long width;
+    public long height;
     public String srcUrl;
     public String style;
     public String ext;
