+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/List.java
@@ -2,12 +2,12 @@ package name.sitorhy.server.model.my.playlists;
 
 public class List{
     public Object dissid;
-    public int dirid;
+    public long dirid;
     public String picurl;
     public String title;
     public String subtitle;
-    public int icontype;
+    public long icontype;
     public String iconurl;
-    public int isshow;
-    public int dir_show;
+    public long isshow;
+    public long dir_show;
 }
