+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/Mydiss.java
@@ -3,7 +3,7 @@ package name.sitorhy.server.model.my.playlists;
 import java.util.ArrayList;
 
 public class Mydiss{
-    public int num;
+    public long num;
     public String title;
     public String laypic;
     public String jumpurl;
