+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/Response.java
@@ -1,8 +1,8 @@
 package name.sitorhy.server.model.my.playlists;
 
 public class Response {
-    public int code;
-    public int subcode;
+    public long code;
+    public long subcode;
     public String msg;
     public Data data;
 }
