+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/Singerinfo.java
@@ -1,5 +1,5 @@
 package name.sitorhy.server.model.my.playlists;
 
 public class Singerinfo{
-    public int singerid;
+    public long singerid;
 }
