+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/my/playlists/Typeinfo.java
@@ -1,7 +1,7 @@
 package name.sitorhy.server.model.my.playlists;
 
 public class Typeinfo{
-    public int type;
+    public long type;
     public String jumpurl;
     public CfinfoBykey cfinfo_bykey;
     public String jumpkey;
