+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/playlist/songs/Preview.java
@@ -1,7 +1,7 @@
 package name.sitorhy.server.model.playlist.songs;
 
 public class Preview{
-    public int trybegin;
-    public int tryend;
-    public int trysize;
+    public long trybegin;
+    public long tryend;
+    public long trysize;
 }
