+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/albums/Company.java
@@ -1,6 +1,6 @@
 package name.sitorhy.server.model.singer.albums;
 
 public class Company{
-    public int companyid;
+    public long companyid;
     public String company_name;
 }
