+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/albums/Data.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 
 public class Data{
     public ArrayList<List> list;
-    public int singer_id;
+    public long singer_id;
     public String singer_mid;
     public String singer_name;
-    public int total;
+    public long total;
 }
