+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/albums/LatestSong.java
@@ -1,7 +1,7 @@
 package name.sitorhy.server.model.singer.albums;
 
 public class LatestSong{
-    public int song_count;
-    public int songid;
+    public long song_count;
+    public long songid;
     public String track_name;
 }
