+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/albums/List.java
@@ -3,7 +3,7 @@ package name.sitorhy.server.model.singer.albums;
 import java.util.ArrayList;
 
 public class List{
-    public int albumid;
+    public long albumid;
     public String album_mid;
     public String album_name;
     public String albumtype;
@@ -11,14 +11,14 @@ public class List{
     public String desc;
     public String attribute5;
     public String ftype;
-    public int type;
+    public long type;
     public String lan;
     public LatestSong latest_song;
-    public int listen_count;
+    public long listen_count;
     public String pub_time;
     public String score;
-    public int shoufa;
-    public int singer_id;
+    public long shoufa;
+    public long singer_id;
     public String singer_mid;
     public String singer_name;
     public ArrayList<Singer> singers;
