+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/albums/Singer.java
@@ -1,7 +1,7 @@
 package name.sitorhy.server.model.singer.albums;
 
 public class Singer{
-    public int singer_id;
+    public long singer_id;
     public String singer_mid;
     public String singer_name;
 }
