+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/albums/SingerAlbum.java
@@ -1,6 +1,6 @@
 package name.sitorhy.server.model.singer.albums;
 
 public class SingerAlbum{
-    public int code;
+    public long code;
     public Data data;
 }
