+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/top/songs/Extra.java
@@ -1,8 +1,8 @@
 package name.sitorhy.server.model.singer.top.songs;
 
 public class Extra{
-    public int listen_count;
+    public long listen_count;
     public String upload_time;
-    public int is_only;
-    public int is_new;
+    public long is_only;
+    public long is_new;
 }
