+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/top/songs/Mv.java
@@ -1,9 +1,9 @@
 package name.sitorhy.server.model.singer.top.songs;
 
 public class Mv{
-    public int id;
+    public long id;
     public String vid;
     public String name;
     public String title;
-    public int vt;
+    public long vt;
 }
