+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/top/songs/Singer.java
@@ -1,12 +1,12 @@
 package name.sitorhy.server.model.singer.top.songs;
 
 public class Singer{
-    public int code;
+    public long code;
     public Data data;
-    public int id;
+    public long id;
     public String mid;
     public String name;
     public String title;
-    public int type;
-    public int uin;
+    public long type;
+    public long uin;
 }
