+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/singer/top/songs/SingerInfo.java
@@ -1,12 +1,12 @@
 package name.sitorhy.server.model.singer.top.songs;
 
 public class SingerInfo{
-    public int area;
-    public int attribute4;
-    public int genre;
+    public long area;
+    public long attribute4;
+    public long genre;
     public String other_name;
-    public int id;
+    public long id;
     public String mid;
     public String name;
-    public int fans;
+    public long fans;
 }
