+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/song/lyric/Response.java
@@ -1,9 +1,9 @@
 package name.sitorhy.server.model.song.lyric;
 
 public class Response {
-    public int retcode;
-    public int code;
-    public int subcode;
+    public long retcode;
+    public long code;
+    public long subcode;
     public String lyric;
     public String trans;
 }
