+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/song/source/Data.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 
 public class Data{
     public String uin;
-    public int retcode;
-    public int verify_type;
+    public long retcode;
+    public long verify_type;
     public String login_key;
     public String msg;
     public ArrayList<String> sip;
@@ -14,5 +14,5 @@ public class Data{
     public String testfilewifi;
     public ArrayList<Midurlinfo> midurlinfo;
     public String servercheck;
-    public int expiration;
+    public long expiration;
 }
