+++ embedded-server-lib/src/main/java/name/sitorhy/server/model/song/source/Req.java
@@ -1,6 +1,6 @@
 package name.sitorhy.server.model.song.source;
 
 public class Req{
-    public int code;
+    public long code;
     public Data data;
 }
